+++ engine/batchstreamengine/engine_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/arquivei/goduck"
 	"github.com/arquivei/goduck/engine/batchstreamengine"
 	"github.com/arquivei/goduck/impl/implprocessor"
-	"github.com/arquivei/goduck/impl/implstream"
+	"github.com/arquivei/goduck/impl/implstream/mockstream"
 
 	"github.com/arquivei/foundationkit/errors"
 	"github.com/stretchr/testify/assert"
@@ -20,7 +20,7 @@ func TestStream(t *testing.T) {
 	processor := implprocessor.New(nil)
 	streams := make([]goduck.Stream, nWorkers)
 	for i := 0; i < nWorkers; i++ {
-		streams[i] = implstream.NewDefaultStream(i, 100)
+		streams[i] = mockstream.NewDefaultStream(i, 100)
 	}
 	defer func() {
 		for _, stream := range streams {
@@ -40,7 +40,7 @@ func TestStreamNoTimeout(t *testing.T) {
 	processor := implprocessor.New(nil)
 	streams := make([]goduck.Stream, nWorkers)
 	for i := 0; i < nWorkers; i++ {
-		streams[i] = implstream.NewDefaultStream(i, 100)
+		streams[i] = mockstream.NewDefaultStream(i, 100)
 	}
 	defer func() {
 		for _, stream := range streams {
@@ -66,7 +66,7 @@ func TestStreamCancel(t *testing.T) {
 	})
 	streams := make([]goduck.Stream, nWorkers)
 	for i := 0; i < nWorkers; i++ {
-		streams[i] = implstream.NewDefaultStream(i, 100)
+		streams[i] = mockstream.NewDefaultStream(i, 100)
 	}
 	defer func() {
 		for _, stream := range streams {
@@ -105,7 +105,7 @@ func TestStreamFatal(t *testing.T) {
 	})
 	streams := make([]goduck.Stream, nWorkers)
 	for i := 0; i < nWorkers; i++ {
-		streams[i] = implstream.NewDefaultStream(i, 100)
+		streams[i] = mockstream.NewDefaultStream(i, 100)
 	}
 	defer func() {
 		for _, stream := range streams {
