+++ impl/implqueue/mockqueue/entity_mockmessage.go
@@ -1,4 +1,4 @@
-package implstream
+package mockqueue
 
 type mockRawMessage struct {
 	data []byte
